backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.7.4
pydantic-settings==2.3.4
python-jose==3.3.0
httpx==0.27.0
redis==5.0.7
celery==5.4.0
python-multipart==0.0.9
